// src/types/twitterCard.ts
export type TwitterCardData = {
  url: string;
  title: string | null;
  description: string | null;
  imageUrl: string | null;
  siteName: string | null;
  siteHandle: string | null;
  siteDomain: string | null;
  cardType: string | null;
  fetchedAt?: string | null;
  themeColor?: string | null;
  faviconUrl?: string | null;
};